prototype/requirements.txt
google-adk>=0.3.0
google-cloud-aiplatform>=1.74.0
firebase-admin>=6.6.0
